+++ packages/components/src/components/link-tile-product/link-tile-product.tsx
@@ -138,7 +138,7 @@ export class LinkTileProduct {
     );
   }
 
-  private onLikeClick = () => {
+  private onLikeClick = (): void => {
     this.like.emit({
       liked: this.liked,
     });