+++ packages/components/src/utils/typography/index.ts
@@ -1,2 +1,3 @@
 export * from './isSizeInherit';
 export * from './typography-align';
+export * from './heading-size';